requirements.txt
discord.py[voice]
asyncpraw
python-dotenv
pynacl
aiosqlite
fuzzywuzzy
python-Levenshtein
aiohttp
cachetools>=5.0
watchdog>=2.0
PyYAML>=6.0
yt-dlp>=2023.7.6
gTTS>=2.3.0
deep-translator>=1.9.0
curl-cffi>=0.5.10
edge-tts>=6.1.9
quart>=0.19.0
quart-discord>=2.0.0
hypercorn>=0.14.0